Switch from tarpaulin to llvm-cov for codecov
Tarpaulin was really slow, not available for Windows/Mac OS,
and had false positives. llvm-cov seems be a good solution.
However we did need to add some tests for derives,
which seems unnecessary, but it's not a big issue.
We have 100% Line Coverage, but to get 100% Region Coverage we would
need to rewrite the DWZ probability table, as it is not currently
possible to get invalid values with .get(), I think.